Moved delete commands to separate file
